+++ pytest_services/locks.py
@@ -2,7 +2,9 @@
 import contextlib
 import json
 import os
+from random import random
 import socket
+import time
 
 import pytest
 import zc.lockfile
@@ -80,23 +82,39 @@ def unlock_display(display, lock_dir, services_log):
     return unlock_resource('display', display, lock_dir, services_log)
 
 
-def lock_resource(name, resource_getter, lock_dir, services_log):
+@pytest.fixture(scope='session')
+def lock_resource_timeout():
+    """Max number of seconds to obtain the lock."""
+    return 20
+
+
+def lock_resource(name, resource_getter, lock_dir, services_log, lock_resource_timeout):
     """Issue a lock for given resource."""
-    with locked_resources(name, lock_dir) as bound_resources:
-        services_log.debug('bound_resources {0}: {1}'.format(name, bound_resources))
+    total_seconds_slept = 0
+    while True:
+        try:
+            with locked_resources(name, lock_dir) as bound_resources:
+                services_log.debug('bound_resources {0}: {1}'.format(name, bound_resources))
+                resource = resource_getter(bound_resources)
+                while resource in bound_resources:
+                    # resource is already taken by someone, retry
+                    services_log.debug('bound resources {0}: {1}'.format(name, bound_resources))
+                    resource = resource_getter(bound_resources)
+                services_log.debug('free resource choosen {0}: {1}'.format(name, resource))
+                bound_resources.append(resource)
+                services_log.debug('bound resources {0}: {1}'.format(name, bound_resources))
+                return resource
+        except zc.lockfile.LockError as err:
+            if total_seconds_slept >= lock_resource_timeout:
+                raise err
+            services_log.debug('lock resource failed: {0}'.format(err))
 
-        resource = resource_getter(bound_resources)
-        while resource in bound_resources:
-            # resource is already taken by someone, retry
-            services_log.debug('bound resources {0}: {1}'.format(name, bound_resources))
-            resource = resource_getter(bound_resources)
-        services_log.debug('free resource choosen {0}: {1}'.format(name, resource))
-        bound_resources.append(resource)
-        services_log.debug('bound resources {0}: {1}'.format(name, bound_resources))
-        return resource
+        seconds_to_sleep = random() * 0.1 + 0.05
+        total_seconds_slept += seconds_to_sleep
+        time.sleep(seconds_to_sleep)
 
 
-def get_free_port(lock_dir, services_log):
+def get_free_port(lock_dir, services_log, lock_resource_timeout):
     """Get free port to listen on."""
     def get_port(bound_resources):
         if bound_resources:
@@ -114,10 +132,10 @@ def get_free_port(lock_dir, services_log):
                 pass
             port += 1
 
-    return lock_resource('port', get_port, lock_dir, services_log)
+    return lock_resource('port', get_port, lock_dir, services_log, lock_resource_timeout)
 
 
-def get_free_display(lock_dir, services_log):
+def get_free_display(lock_dir, services_log, lock_resource_timeout):
     """Get free display to listen on."""
     def get_display(bound_resources):
         display = 100
@@ -129,15 +147,15 @@ def get_free_display(lock_dir, services_log):
                 continue
             return display
 
-    return lock_resource('display', get_display, lock_dir, services_log)
+    return lock_resource('display', get_display, lock_dir, services_log, lock_resource_timeout)
 
 
 @pytest.fixture(scope='session')
-def port_getter(request, lock_dir, services_log):
+def port_getter(request, lock_dir, services_log, lock_resource_timeout):
     """Lock getter function."""
     def get_port():
         """Lock a free port and unlock it on finalizer."""
-        port = get_free_port(lock_dir, services_log)
+        port = get_free_port(lock_dir, services_log, lock_resource_timeout)
 
         def finalize():
             unlock_port(port, lock_dir, services_log)
@@ -147,11 +165,11 @@ def port_getter(request, lock_dir, services_log):
 
 
 @pytest.fixture(scope='session')
-def display_getter(request, lock_dir, services_log):
+def display_getter(request, lock_dir, services_log, lock_resource_timeout):
     """Display getter function."""
     def get_display():
         """Lock a free display and unlock it on finalizer."""
-        display = get_free_display(lock_dir, services_log)
+        display = get_free_display(lock_dir, services_log, lock_resource_timeout)
         request.addfinalizer(lambda: unlock_display(display, lock_dir, services_log))
         return display
     return get_display